beeline database update to create

+++ app/src/main/java/com/wenwanggarzagao/beeline/data/Location.java
@@ -6,7 +6,7 @@ package com.wenwanggarzagao.beeline.data;
 public class Location {
 
     // TODO what makes sense as args to Location?
-    public Location(String address, String city, String state, short zip) {
+    public Location(String address, String city, String state, int zip) {
         this.address = address;
         this.city = city;
         this.state = state;
@@ -16,7 +16,7 @@ public class Location {
     public String address;
     public String city;
     public String state;
-    public short zip;
+    public int zip;
 
 
     @Override
